# src/main/resources/swagger/api.yml
...
openapi: 3.0.1
info:
  title: thg-procure-questionnaire-service
  description: This is to serve as the documentation of the questionnaire-service APIs for THG Procure.
  version: 1.0.0
  contact:
    name: Abhijeet Singh
    email: abhijeet.singh@cloudesign.com
servers:
  - url: https://dev.qts.thg-procure.com
  - url: https://uat.qts.thg-procure.com
  - url: https://qts.thg-procure.com
paths:
  /questionnaire:
    description: API endpoints to manage questionnaires
    post:
      description: Add a new questionnaire
      parameters:
        - in: header
          name: qts-client-id
          description: To identify client of the questionnaire service
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createQuestionnaire'
      responses:
        200:
          description: Questionnaire created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getQuestionnaire'
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        401:
          description: Unauthorized
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        400:
          description: Bad Request
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        404:
          description: Client Not Found
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
  /questionnaires:
    description: API endpoints to fetch list of questionnaires
    get:
      description: Get list of questionnaires
      parameters:
        - in: header
          name: qts-client-id
          description: To identify client of the questionnaire service
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: query
          name: name
          description: Filter questionnaire list on basis of names
          required: false
          schema:
            type: string
            example: SupplierOnboardingForm
        - in: query
          name: description
          description: Filter questionnaire list on basis of description
          required: false
          schema:
            type: string
            example: SupplierOnboardingForm
      responses:
        200:
          description: Questionnaires fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/getQuestionnaires'
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        401:
          description: Unauthorized
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        400:
          description: Bad Request
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        404:
          description: Client Not Found
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
  /questionnaire/{id}:
    description: API endpoints to manage a questionnaire
    put:
      description: Update an existing questionnaire
      parameters:
        - in: header
          name: qts-client-id
          description: To identify client of the questionnaire service
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: path
          name: id
          description: The unique id of the questionnaire
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateQuestionnaire'
      responses:
        200:
          description: Questionnaire updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getQuestionnaire'
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        404:
          description: Requested Questionnaire id not found
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        401:
          description: Unauthorized
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        400:
          description: Bad Request
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
    get:
      description: Get details of a questionnaire by ID
      parameters:
        - in: header
          name: qts-client-id
          description: To identify client of the questionnaire service
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: path
          name: id
          description: The unique id of the questionnaire
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      responses:
        200:
          description: Questionnaire fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getQuestionnaire'
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        404:
          description: Requested Questionnaire id not found
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        401:
          description: Unauthorized
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        400:
          description: Bad Request
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
    delete:
      description: Delete a specific questionnaire
      parameters:
        - in: header
          name: qts-client-id
          description: To identify client of the questionnaire service
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: path
          name: id
          description: The unique id of the questionnaire to delete
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      responses:
        204:
          description: Questionnaire deleted successfully
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        404:
          description: Requested Questionnaire id not found
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        401:
          description: Unauthorized
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
        400:
          description: Bad Request
          headers:
            qtsClientId:
              schema:
                type: string
                format: uuid
              description: Requested Client Id
...
components:
  schemas:
    createQuestionnaire:
      type: object
      required: [name, description, schema]
      properties:
        name:
          type: string
        description:
          type: string
        schema:
          type: string
    getQuestionnaires:
      type: object
      required: [id, name, description, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedBy:
          type: string
          format: uuid
    getQuestionnaire:
      type: object
      required: [id, name, description, schema, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        schema:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedBy:
          type: string
          format: uuid
    updateQuestionnaire:
      type: object
      required: [questionnaireName, questionnaireDescription, questionnaireSchema]
      properties:
        name:
          type: string
        description:
          type: string
        schema:
          type: string
...